refactor(data-context): move QueryExecutionService from Hub to Core
Move IQueryExecutionService and QueryExecutionService from Bolt.Hub.Services
to XFramework.Core.DataContext so any service module can use them directly
without depending on the Hub.
